stock-portfolio-project/backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.61.1
frozendict==2.4.7
idna==3.11
ImageIO==2.37.2
joblib==1.5.3
kiwisolver==1.4.9
kombu==5.6.2
lazy_loader==0.4
matplotlib==3.10.8
multitasking==0.0.12
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
patsy==1.0.2
peewee==4.0.1
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==7.34.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
redis==7.2.1
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
statsmodels==0.14.6
threadpoolctl==3.6.0
tensorflow==2.20.0
tifffile==2026.3.3
tvdatafeed @ git+https://github.com/rongardF/tvdatafeed.git@e6f6aaa7de439ac6e454d9b26d2760ded8dc4923
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
yfinance==1.2.0



